LayoutTests/imported/w3c/web-platform-tests/permissions-policy/private-state-token-issue-allowed-by-permissions-policy-attribute.tentative.https.sub-expected.txt



FAIL Permissions policy "private-state-token-issuance" can be enabled in same-origin iframe using allow="private-state-token-issuance" attribute assert_equals: Private State Token issuance request expected 2 but got 1
FAIL Permissions policy "private-state-token-issuance" can be enabled in cross-origin iframe using allow="private-state-token-issuance" attribute assert_equals: Private State Token issuance request expected 2 but got 1
